requirements.txt
duckdb==1.1.0
neo4j-rust-ext==5.24.0
python-dotenv==1.0.1
pandas==2.2.3